requirements.txt
pyserial
numpy
scipy
joblib
scikit-learn
keyboard
